requirements.txt
fastapi
uvicorn
gunicorn
pydantic
pydantic-settings
python-jose[cryptography]
google-genai
openai
httpx